+++ corehq/ex-submodules/casexml/apps/case/dbaccessors/__init__.py
@@ -1,2 +1,3 @@
 from .all_cases import get_all_case_owner_ids
 from .by_owner import get_open_case_ids, get_closed_case_ids
+from .related import get_indexed_case_ids, get_reverse_indexed_case_ids
